backend/requirements.txt
fastapi
uvicorn
python-dotenv
resend
supabase
requests
email-validator
resend